% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_new_sites.R
\name{adjust_new_sites}
\alias{adjust_new_sites}
\title{Adjust New Sites' Sampling Weights}
\usage{
adjust_new_sites(sampling_weight, samples, new_venue_factor)
}
\arguments{
\item{sampling_weight}{The name of the variable with the sampling weights.}

\item{samples}{The name of the variable containing the samples the venue has been eligible for.}

\item{new_venue_factor}{The scaling factor for new venues.}
}
\value{
A vector of adjusted sampling weights.
}
\description{
Increases the sampling weights of new venues. New venues are ones have not
been eligible to be selected in a previous sample.
}
\examples{
\dontrun{
venues \%>\%
  mutate(sampling_weight = adjust_new_sites(sampling_weight, samples, 2))
}
}
