% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_airtable_records.R
\name{chunk_data}
\alias{chunk_data}
\title{Split Data Frame into Chunks}
\usage{
chunk_data(data, chunk_size)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{chunk_size}{The number of rows in each chunk.}
}
\value{
A list of data frames or tibbles.
}
\description{
Breaks a data frame into equal sized chunks for upload to Airtable.
}
\keyword{internal}
