% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{connect_to_airtable}
\alias{connect_to_airtable}
\title{Connect to the Airtable API}
\usage{
connect_to_airtable(
  base = "appOLCptG2wxvoGtH",
  tables = c("Venues", "Samples", "Visit Check-Out")
)
}
\arguments{
\item{base}{ID for the airtable base.}

\item{tables}{Names of the tables to include.}
}
\value{
An Airtable base object from \code{\link[airtabler:airtable]{airtabler::airtable()}}.
}
\description{
Makes a connection to the Airtable API using the \code{\link[airtabler:airtabler-package]{airtabler::airtabler()}}
package.
}
\details{
Expects an API key in an environment variable \code{AIRTABLE_API_KEY}. To start R
session with the initialized environment variable create an .Renviron file
with a line like this: \verb{AIRTABLE_API_KEY=<YOUR API KEY>}.

See https://airtable.com/appOLCptG2wxvoGtH/api/docs#curl/authentication for
details on obtaining an Airtable API key.
}
