% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata.R
\name{default_cleaning_do_file}
\alias{default_cleaning_do_file}
\title{Provides path to included Stata data cleaning code.}
\usage{
default_cleaning_do_file()
}
\value{
Path to cleaning do file.
}
\description{
Provides path to included Stata data cleaning code.
}
\keyword{internal}
