% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample.R
\name{draw_sample}
\alias{draw_sample}
\title{Draw Venue Sample}
\usage{
draw_sample(data, weights, n)
}
\arguments{
\item{data}{Data frame with venues.}

\item{weights}{Name of the weighting variable.}

\item{n}{Either a single number of venues to select, or a data frame with
stratification plan (see below).}
}
\value{
The venue data frame with the variable \code{sampled} added, which
indicates whether the venue was selected.
}
\description{
Draws a weighted, without replacement sample of venues. Supports both
stratified and unstratified samples.
}
\section{Stratification Plan}{

To draw a stratified sample, include a data frame with the stratification
plan.

The data frame should include a row for each stratum with the values of
the stratification variables and the number of venues to sample. The
stratification variables must have the same name and coding as in the
venue data. The variable with the the number of venues to sample must be
named \code{n}.

For example:\tabular{lll}{
   \strong{dv_shelter} \tab \strong{zone} \tab \strong{n} \cr
   TRUE \tab Eastern Placer \tab 1 \cr
   TRUE \tab Central Placer \tab 2 \cr
   TRUE \tab South Placer \tab 3 \cr
   FALSE \tab Eastern Placer \tab 4 \cr
   FALSE \tab Central Placer \tab 5 \cr
   FALSE \tab South Placer \tab 6 \cr
}


This can be generated with the following R code:\preformatted{tibble::tribble(
  ~dv_shelter, ~zone           , ~n,
   TRUE      , "Eastern Placer",  1,
   TRUE      , "Central Placer",  2,
   TRUE      , "South Placer"  ,  3,
   FALSE     , "Eastern Placer",  4,
   FALSE     , "Central Placer",  5,
   FALSE     , "South Placer"  ,  6
)
}
}

\examples{
\dontrun{
  draw_sample(venues, sampling_weight, 5)

  strata = tibble::tribble(
    ~zone          , ~n,
    "Eastern Placer",  1,
    "Central Placer",  3,
    "South Placer"  ,  5
  )
  draw_sample(venues, sampling_weight, strata)
}
}
