% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{fetch_redcap_data}
\alias{fetch_redcap_data}
\title{Downloads REDCap Data}
\usage{
fetch_redcap_data(
  ...,
  redcap_uri = redcap_api_url(),
  token = redcap_token(),
  verbose = FALSE
)
}
\arguments{
\item{...}{Options passed to \code{\link[REDCapR:redcap_read]{REDCapR::redcap_read()}}.}

\item{redcap_uri}{REDCap API URL. Defaults to \code{\link[=redcap_api_url]{redcap_api_url()}}.}

\item{token}{REDCap API Token. Defaults to main project token. Use RDS
project token with \code{redcap_token("rds")}.}

\item{verbose}{Should messages be printed to the R console during the
operation. The verbose output might contain sensitive information (e.g.
PHI), so turn this off if the output might be visible somewhere public.}
}
\value{
A tibble with the requested data.
}
\description{
Downloads data from REDCap by wrapping \code{\link[REDCapR:redcap_read]{REDCapR::redcap_read()}}. The API URL
is filled automatically, but can be overridden. The token for the main
project is also filled automatically, but can be overridden or replaced with
the RDS project token.
}
\seealso{
\code{\link[=redcap_token]{redcap_token()}}
}
