% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stata_date.R
\name{format_stata_date}
\alias{format_stata_date}
\title{Convert a Stata date to text}
\usage{
format_stata_date(date, format = "\%B \%e, \%Y")
}
\arguments{
\item{date}{Stata date.}

\item{format}{Format string for character representation. See
\code{\link[base:strptime]{base::strptime()}} for options. Defaults to formatting like January 1,
2022.}
}
\value{
A character string with the formatted date.
}
\description{
Converts a Stata date to a character string.
}
