% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_airtable_records.R
\name{insert_airtable_records}
\alias{insert_airtable_records}
\title{Upload Data to Airtable}
\usage{
insert_airtable_records(data, base, table, chunk_size = 10)
}
\arguments{
\item{data}{Data frame of properly named and formatted data.}

\item{base}{Airtable Base ID.}

\item{table}{Name of the Airtable table.}

\item{chunk_size}{Number of rows in each data chunk. Airtable accepts a
maximum of 10 rows in one request, so this cannot be greater than 10.
Default is 10.}
}
\description{
Uploads a properly named and formatted data frame to Airtable. It handles
converting the data frame to JSON that Airtable understands, and splitting
the data into chunks of no more than 10 rows to comply with Airtable limits.
}
\examples{
\dontrun{
  insert_airtable_records(
    data_for_airtable,
    "appOLCptG2wxvoGtH",
    "Samples"
  )
}
}
