% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_airtable_records.R
\name{json_for_airtable}
\alias{json_for_airtable}
\title{Convert Data Frame to Airtable JSON}
\usage{
json_for_airtable(data)
}
\arguments{
\item{data}{Data frame/tibble with properly named and formatted columns.}
}
\description{
Converts a data frame or tibble to the JSON data format accepted by the
Airtable API.
}
\keyword{internal}
