% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_sheltered_unsheltered.R
\name{observed_sheltered_share}
\alias{observed_sheltered_share}
\title{Calculate the Sheltered Share in Venue Data}
\usage{
observed_sheltered_share(sheltered_indicator, peh_count)
}
\arguments{
\item{sheltered_indicator}{A character vector/variable with the
sheltered/unsheltered indicator.}

\item{peh_count}{A numeric vector/variable with the PEH estimates.}
}
\description{
Calculate the share of PEH who are sheltered in the venue data. The sheltered
share is \verb{sheltered PEH / (sheltered PEH + unsheltered PEH)}.
}
\examples{
sheltered_indicator = c("Sheltered", "Unsheltered", NA)
peh_count = c(5, 10, 20)
observed_sheltered_share(sheltered_indicator, peh_count)
}
