% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{post_to_airtable}
\alias{post_to_airtable}
\title{Send Upload Request to Airtable}
\usage{
post_to_airtable(data, base, table)
}
\arguments{
\item{data}{A data frame or tibble with no more than 10 rows.}

\item{base}{Airtable Base ID.}

\item{table}{Airtable table name.}
}
\description{
Sends an HTTP POST request to Airtable. It handles converting the data frame
to JSON, and creating and submitting the request.
}
\keyword{internal}
