% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sample_for_airtable.R
\name{prepare_sample_for_airtable}
\alias{prepare_sample_for_airtable}
\title{Prepare sampling data for Airtable}
\usage{
prepare_sample_for_airtable(
  sample,
  venue_id,
  sampling_weight,
  sampled_indicator,
  size_category,
  county,
  zone,
  round,
  survey_target_mapping,
  staffing_mapping
)
}
\arguments{
\item{sample}{Data frame with the full sample (ie. both selected and not
selected sites).}

\item{venue_id}{Name of the column with the internal Airtable venue ID (different than
the venue_id we commonly use like placer_123).}

\item{sampling_weight}{Name of the column with the sampling weights.}

\item{sampled_indicator}{Name of the column which indicates whether the
venue was selected.}

\item{size_category}{Name of the column with the venue size category.}

\item{county}{Character string with the county name of the sample.}

\item{zone}{Character string with the zone of the sample.}

\item{round}{Character string or number with the round of the sample.}

\item{survey_target_mapping}{Named vector with the size categories as the names and their corresponding survey targets as the values.}

\item{staffing_mapping}{Named vector with the size categories as the names and their corresponding staffing levels as the values.}
}
\value{
A data frame ready to be passed to \code{\link[=insert_airtable_records]{insert_airtable_records()}}.
}
\description{
Formats the sampling data for submission to Airtable. It selects only the
variables that will be sent to Airtable (internal Airtable venue id, sampling
weight, & selected indicator) and adds the sample metadata (ie. county, zone,
and round).
}
\examples{
\dontrun{
  prepare_sample_for_airtable(
    sample = final_sample,
    venue_id = id,
    sampling_weight = sampling_weight,
    sampled_indicator = sampled,
    size_category = size_category,
    county = "Sonoma",
    zone = "Santa Rosa",
    round = 2,
    survey_target_mapping = c(
      "[5,25)" = "4-5",
      "[25,50)" = "8-10",
      "[50,75)" = "15",
      "[75,150)" = "30",
      "[150,Inf)" = "50-60"
    ),
    staffing_mapping = c(
      "[5,25)" = "quarter of the team, half day",
      "[25,50)" = "half of the team, half day",
      "[50,75)" = "half of the team, full day or full team, half day",
      "[75,150)" = "full team, 1 day",
      "[150,Inf)" = "full team, 2 days"
    )
  )
}
}
