% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_sheltered_unsheltered.R
\name{sheltered_or_unsheltered}
\alias{sheltered_or_unsheltered}
\title{Create Sheltered or Unsheltered Indicator}
\usage{
sheltered_or_unsheltered(
  site_category,
  sheltered_categories = c("Emergency Shelter"),
  unsheltered_categories = c("Encampment")
)
}
\arguments{
\item{site_category}{Character vector with site categories.}

\item{sheltered_categories}{Character vector with the site categories that
represent sheltered venues.}

\item{unsheltered_categories}{Character vector with the site categories that
represent unsheltered venues.}
}
\value{
A character vector with either "Sheltered", "Unsheltered, or NA.
}
\description{
Based on the site category determine whether a venue counts as sheltered or
unsheltered (or neither).
}
\examples{
site_category = c(
  "Emergency Shelter", "Encampment",
  "Non Shelter Venue", "Hotspot"
)
sheltered_or_unsheltered(site_category)
}
