% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_sheltered_unsheltered.R
\name{adjust_sheltered_unsheltered}
\alias{adjust_sheltered_unsheltered}
\title{Adjust PEH Counts for Sheltered/Unsheltered Share}
\usage{
adjust_sheltered_unsheltered(peh_count, site_category, sheltered_share)
}
\arguments{
\item{peh_count}{A numeric vector of venue PEH counts.}

\item{site_category}{A character vector of venue site categories.}

\item{sheltered_share}{The target sheltered share as a single numeric.}
}
\value{
A vector of adjusted PEH counts.
}
\description{
Makes the adjustment to sampling weights so that the weighted average share
of residential venue (ie. emergency shelters or encampments) PEH who are
sheltered matches a target (usually from PIT counts). It's designed for use
within a \code{\link[dplyr:mutate]{dplyr::mutate()}} pipeline when calculating the sampling weights.
}
\examples{
\dontrun{
  venue_data \%>\%
    mutate(
      sampling_weight = adjust_sheltered_unsheltered(
        final_peh_estimate,
        site_category,
        0.45
      )
    )
}
}
