% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_sheltered_unsheltered.R
\name{adjust_sheltered_unsheltered}
\alias{adjust_sheltered_unsheltered}
\alias{sheltered_or_unsheltered}
\alias{calculate_adjusters}
\alias{adjust_count}
\title{Adjust PEH Counts for Sheltered/Unsheltered Share}
\usage{
adjust_sheltered_unsheltered(sampling_weight, site_category, sheltered_share)

sheltered_or_unsheltered(
  site_category,
  sheltered_categories = c("Emergency Shelter"),
  unsheltered_categories = c("Encampment")
)

calculate_adjusters(sheltered_indicator, sampling_weight, sheltered_share)

adjust_count(sheltered_indicator, sampling_weight, adjusters)
}
\arguments{
\item{sampling_weight}{A numeric vector/variable with the current sampling
weights.}

\item{site_category}{A character vector/variable of venue site categories.}

\item{sheltered_share}{The target sheltered share as a single numeric.}

\item{sheltered_categories}{Character vector with the site categories that
represent sheltered venues.}

\item{unsheltered_categories}{Character vector with the site categories that
represent unsheltered venues.}

\item{adjusters}{A numeric value with the adjuster.}
}
\value{
A vector of adjusted sampling weights.
}
\description{
Makes the adjustment to sampling weights so that the weighted average share
of residential venue (ie. emergency shelters or encampments) PEH who are
sheltered matches a target (usually from PIT counts). It's designed for use
within a \code{\link[dplyr:mutate]{dplyr::mutate()}} pipeline when calculating the sampling weights or
the component functions (listed below) can be used to customize the
adjustment.
}
\section{Functions}{
\itemize{
\item \code{sheltered_or_unsheltered}: Create Sheltered or Unsheltered
Indicator

Based on the site category determine whether a venue counts as sheltered or
unsheltered (or neither). Returns a character vector with either
"Sheltered", "Unsheltered, or NA.

\item \code{calculate_adjusters}: Calculate the adjustment factors for
the sheltered and unsheltered PEH estimates.

\item \code{adjust_count}: Make Sheltered/Unsheltered
Adjustment
}}

\examples{
\dontrun{
  venue_data \%>\%
    mutate(
      sampling_weight = adjust_sheltered_unsheltered(
        final_peh_estimate,
        site_category,
        0.45
      )
    )}

site_category = c(
  "Emergency Shelter", "Encampment",
  "Non Shelter Venue", "Hotspot"
)
sheltered_indicator = sheltered_or_unsheltered(site_category)

sampling_weight = c(5, 10, 20, 15)
calculate_adjusters(
  sheltered_indicator,
  sampling_weight,
  0.33
)

adjust_count(
  sheltered_indicator,
  sampling_weight,
  c(sheltered = 0.5, unsheltered = 1.5)
)
}
