% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_from_airtable.R
\name{fix_names}
\alias{fix_names}
\title{Fix Variable Names}
\usage{
fix_names(name)
}
\arguments{
\item{name}{Variable name as a string.}
}
\description{
Make variable names lower case and switch spaces to underscores.
}
\keyword{internal}
