% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sample_for_airtable.R
\name{prepare_sample_for_airtable}
\alias{prepare_sample_for_airtable}
\title{Prepare sampling data for Airtable}
\usage{
prepare_sample_for_airtable(
  sample,
  venue_id,
  sampling_weight,
  sampled_indicator,
  sample_county,
  sample_zone,
  sample_round
)
}
\arguments{
\item{sample}{Data frame with the full sample (ie. both selected and not
selected sites).}

\item{venue_id}{Name of the column with the internal Airtable venue ID (different than
the venue_id we commonly use like placer_123).}

\item{sampling_weight}{Name of the column with the sampling weights.}

\item{sampled_indicator}{Name of the column which indicates whether the
venue was selected.}

\item{sample_county}{Character string with the county name of the sample.}

\item{sample_zone}{Character string with the zone of the sample.}

\item{sample_round}{Character string or number with the round of the sample.}
}
\value{
A data frame ready to be passed to \code{\link[=insert_airtable_records]{insert_airtable_records()}}.
}
\description{
Formats the sampling data for submission to Airtable. It selects only the
variables that will be sent to Airtable (internal Airtable venue id, sampling
weight, & selected indicator) and adds the sample metadata (ie. county, zone,
and round).
}
\examples{
\dontrun{
  prepare_sample_for_airtable(
    sample = final_sample,
    venue_id = id,
    sampling_weight = sampling_weight,
    sampled_indicator = sampled,
    county = "Sonoma",
    zone = "Santa Rosa",
    round = 2
  )
}
}
